1팀끝

+++ src/main/java/com/nhnacademy/bankservice/Money.java
@@ -18,7 +18,7 @@ public class Money {
         return new Money(money, Currency.DOLLAR);
     }
 
-    static Money won(long money) {
+    static Money won(double money) {
         return new Money(money, Currency.WON);
     }
